backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
chromadb
pydantic
requests
